requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pyPDF2
chromadb
pdf2image
faiss-cpu
langchain_google_genai